Further DRY-ing up of code. WIP
This is pretty close, but the one down side of this is that curlget is
now also called when creating a new todo and this is a redundant call.
Unfortunately I have not figured out how to make this call
conditionally. Simply including these statements within a do statement
in the curlpostfields conditional doesn't work.