Perform curlget conditionally within curlpost
Previous commits DRY-ed up the code so all TeuxDeux functions used
a shared a curlpost function, but meant that a curlget was performed
even when not necessary. I.e. when adding a new todo it is not necessary
to fetch a list of the todos. Fetching a list of todos is only necessary
when making some change to an existing todo.

I'm quite sure I could make yet more improvements to this code, but it
is a lot better than it was!

With this last change of the last few commits, closes #1.