Use Web.Encodings to URL Encode new todo text
Noticed text was getting chopped off at ampersands, etc. URL encoding
seems to be the correct option. Just passing a "&amp;" is no good. Needs
to be "%26".

I know Web.Encodings is depreciated, but I'm unsure of it's replacement.
If need be I could just pull out the encodeUrl bit from it and include
directly.