Get today and crossoff working with system call approach
As mentioned in the previous commit I decided to give up on
Network.Curl and libcurl. There seem to be issues there that could be
due to either of them two or me. I know that the curl command line works
so I'll just make this program a wrapper around that. It's not as pretty
or elegant, but it'll work.

So this commit is basically that, moving from Network.Curl to using
readprocess. Still much tidying up to do.

- Need to figure out hiding the curl UI feedback
- Need to figure out re-using cookies. Network.Curl did strange things
  with the cookie files, but I know command line curl works properly so
  I can cache the token and re-use cookies, etc to save logging on
  everytime.