Enable putoff and moveto commands
This is an utter mess and soooooo unhaskelly, but nearly back to the
original functionality.

I'm most miserable about the passing of args to curlpost. Ideally I'd
like to pass an association list representing the json data and then
convert this into a string representation of the json data for posting
with curl. Perhaps Text.Json can do this (probably Aeson can do if
Text.Json can't), but I couldn't figure it out so I've done this an
crappy way.

The other thing I need to do is to do is make it so the resposition
moves it to the bottom of the list. Not figured that out yet. Doesn't
happen automatically anymore.

Maybe one day I can make this better.