Add back in "new" todo functionality.
Now, finally, has the same functionality it did four years ago!

But is a messy piece of crap. In fact it's pretty horrible, but it works
so if I'm really honest I'm probably just going to leave it like this. I
just can gel with Haskell like I can Erlang (and I ain't saying I'm good
there).

Things of note:

- Remembered why I needed Just/Maybe on the curlpost
- The json bit in the curlpost, need to let the statements rather than
  returning them directly. I don't understand why, but get errors.
- The way the curlpost works to sort out the correct json string
  depending on whether adding a new todo or modifying an existing is
  just horrible. I really need to figure out using association lists as
  args and cleanly converting these to a json string.